requirements.txt
Flask==3.0.2
GitPython==3.1.41
PyYAML==6.0.1
cryptography==42.0.5
